Split settings into dev and production
There's now two settings files. The production one contains the majority
of the configuration and is named `settings_prod`. The settings for local
development are stored in `settings.py` which will import `settings_prod`
and override some values.

The reason to name the settings file for development `settings.py` is that
`./manage runserver` will automatically pick up that file. For the
production environment we can set the settings file in the main WSGI script.